idf_monitor: Fix console performance
https://github.com/espressif/esp-idf/commit/d1c895e3ed666be6009327dcf68c189b20f8e9a3
introduced filtering options which handles the last (unterminated) line
with a delay. This introduced poor performance for console applications
when the user interacts with the ESP32 device in the same line because
there is an artificial delay for each key-press.

https://github.com/espressif/esp-idf/commit/a13a2f7e34ebb3690488197365eab0ecd415f36e
decreased the delay but there are still delays for each key-presses.

The current fix makes only one delay per line instead of each
key-presses. When an unterminated line is detected then no more
"delayed" print is used for the given line.

Fixes https://github.com/espressif/esp-idf/issues/2054