check if stdin is attached to TTY
idf-monitor uses miniterm from pyserial, which sets terminal properties,
such as noncanonical mode, with termios. This requires stdin to be
attached to terminal, otherwise miniterm exits with the following
exception: "termios.error: (25, 'Inappropriate ioctl for device')"

This adds early check if stdin is attached to TTY and now idf-monitor
exits gracefully if not.

Closes https://github.com/espressif/esp-idf/issues/11150

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>