fix(idf_monitor): don't discard input for all socket:// ports
Previously it was assumed that 'socket://' ports are only used for
testing of idf_monitor. In testing mode, all input from console was
discarded.
This leads to a problem when idf_monitor is used together with QEMU
and QEMU console is attached using a TCP socket. Since console input
is ignored, it's not possible to exit IDF monitor using keyboard
shortcuts.
To make it possible to use idf_monitor with 'socket://' ports, this
commit adds an environment variable, ESP_IDF_MONITOR_TEST. The socket
test mode (i.e. discarding of the console input) will be enabled
in idf_monitor only if this variable is set.
The tests have been updated accordingly.