fix: correctly decode string type in ANSIColorConverter
Currently there is a bug in the converter related to how it processes
multibyte characters. The problem lies in the 'decode_buffer,' where
input bytes are being added multiple times, as we are redundantly
including bytes that are already present in the buffer when the decode
fails.

The ansi_color_converter is also used in idf.py. The difference is that
monitor is writing bytes directly to sys.stdout.buffer, which does not
undergo decode in converter, while idf.py is using sys.stdout, where the
decode is done and this problem manifests.

Closes https://github.com/espressif/esp-idf-monitor/issues/12
Closes https://github.com/espressif/esp-idf/issues/13046

Signed-off-by: Frantisek Hrbata <frantisek.hrbata@espressif.com>